refactor(backend): extract code studio scaffold spec
- Move deterministic Code Studio topic library, concept matching, and spec inference into code_studio_scaffold_spec.py
- Keep code_studio_template_scaffold.py as renderer registry plus compatibility re-export surface
- Update architecture and runtime cleanup audit docs with risk, rollback, and verification notes
